Fix: tcp_trim_head for MPTCP
The head will get trimmed, because part of the data has been acked. By
trimming, we loose the already written MPTCP-option. Additionally, we
know that part of the data reached the destination and thus the
MPTCP-DSS-opt probably reached the destination too.

So, we just unset MPTCPHDR_SEQ here in order to send out a segment
without DSS.

Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>